+++ internal/query/result_range_test.go
@@ -348,7 +348,7 @@ func TestResultRange(t *testing.T) {
 	require.NoError(t, err)
 	defer r.Close(ctx)
 	rsCount := 0
-	for rs, err := range r.Range(ctx) {
+	for rs, err := range r.ResultSets(ctx) {
 		require.NoError(t, err)
 		rowsCount := 0
 		for _, err := range rs.Range(ctx) {