+++ trace/query_gtrace.go
@@ -4,8 +4,6 @@ package trace
 
 import (
 	"context"
-
-	"github.com/ydb-platform/ydb-go-genproto/protos/Ydb_TableStats"
 )
 
 // queryComposeOptions is a holder of options
@@ -755,41 +753,6 @@ func (t *Query) Compose(x *Query, opts ...QueryComposeOption) *Query {
 		}
 	}
 	{
-		h1 := t.OnResultNextPart
-		h2 := x.OnResultNextPart
-		ret.OnResultNextPart = func(q QueryResultNextPartStartInfo) func(QueryResultNextPartDoneInfo) {
-			if options.panicCallback != nil {
-				defer func() {
-					if e := recover(); e != nil {
-						options.panicCallback(e)
-					}
-				}()
-			}
-			var r, r1 func(QueryResultNextPartDoneInfo)
-			if h1 != nil {
-				r = h1(q)
-			}
-			if h2 != nil {
-				r1 = h2(q)
-			}
-			return func(info QueryResultNextPartDoneInfo) {
-				if options.panicCallback != nil {
-					defer func() {
-						if e := recover(); e != nil {
-							options.panicCallback(e)
-						}
-					}()
-				}
-				if r != nil {
-					r(info)
-				}
-				if r1 != nil {
-					r1(info)
-				}
-			}
-		}
-	}
-	{
 		h1 := t.OnResultNextResultSet
 		h2 := x.OnResultNextResultSet
 		ret.OnResultNextResultSet = func(q QueryResultNextResultSetStartInfo) func(QueryResultNextResultSetDoneInfo) {
@@ -1308,21 +1271,6 @@ func (t *Query) onResultNew(q QueryResultNewStartInfo) func(info QueryResultNewD
 	}
 	return res
 }
-func (t *Query) onResultNextPart(q QueryResultNextPartStartInfo) func(info QueryResultNextPartDoneInfo) {
-	fn := t.OnResultNextPart
-	if fn == nil {
-		return func(QueryResultNextPartDoneInfo) {
-			return
-		}
-	}
-	res := fn(q)
-	if res == nil {
-		return func(QueryResultNextPartDoneInfo) {
-			return
-		}
-	}
-	return res
-}
 func (t *Query) onResultNextResultSet(q QueryResultNextResultSetStartInfo) func(info QueryResultNextResultSetDoneInfo) {
 	fn := t.OnResultNextResultSet
 	if fn == nil {
@@ -1413,6 +1361,7 @@ func (t *Query) onRowScanStruct(q QueryRowScanStructStartInfo) func(info QueryRo
 	}
 	return res
 }
+
 // Internals: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#internals
 func QueryOnNew(t *Query, c *context.Context, call call) func() {
 	var p QueryNewStartInfo
@@ -1424,6 +1373,7 @@ func QueryOnNew(t *Query, c *context.Context, call call) func() {
 		res(p)
 	}
 }
+
 // Internals: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#internals
 func QueryOnClose(t *Query, c *context.Context, call call) func(error) {
 	var p QueryCloseStartInfo
@@ -1436,6 +1386,7 @@ func QueryOnClose(t *Query, c *context.Context, call call) func(error) {
 		res(p)
 	}
 }
+
 // Internals: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#internals
 func QueryOnPoolNew(t *Query, c *context.Context, call call) func(limit int) {
 	var p QueryPoolNewStartInfo
@@ -1448,6 +1399,7 @@ func QueryOnPoolNew(t *Query, c *context.Context, call call) func(limit int) {
 		res(p)
 	}
 }
+
 // Internals: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#internals
 func QueryOnPoolClose(t *Query, c *context.Context, call call) func(error) {
 	var p QueryPoolCloseStartInfo
@@ -1460,6 +1412,7 @@ func QueryOnPoolClose(t *Query, c *context.Context, call call) func(error) {
 		res(p)
 	}
 }
+
 // Internals: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#internals
 func QueryOnPoolTry(t *Query, c *context.Context, call call) func(error) {
 	var p QueryPoolTryStartInfo
@@ -1472,6 +1425,7 @@ func QueryOnPoolTry(t *Query, c *context.Context, call call) func(error) {
 		res(p)
 	}
 }
+
 // Internals: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#internals
 func QueryOnPoolWith(t *Query, c *context.Context, call call) func(_ error, attempts int) {
 	var p QueryPoolWithStartInfo
@@ -1485,6 +1439,7 @@ func QueryOnPoolWith(t *Query, c *context.Context, call call) func(_ error, atte
 		res(p)
 	}
 }
+
 // Internals: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#internals
 func QueryOnPoolPut(t *Query, c *context.Context, call call) func(error) {
 	var p QueryPoolPutStartInfo
@@ -1497,6 +1452,7 @@ func QueryOnPoolPut(t *Query, c *context.Context, call call) func(error) {
 		res(p)
 	}
 }
+
 // Internals: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#internals
 func QueryOnPoolGet(t *Query, c *context.Context, call call) func(error) {
 	var p QueryPoolGetStartInfo
@@ -1509,6 +1465,7 @@ func QueryOnPoolGet(t *Query, c *context.Context, call call) func(error) {
 		res(p)
 	}
 }
+
 // Internals: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#internals
 func QueryOnPoolChange(t *Query, limit int, index int, idle int, inUse int) {
 	var p QueryPoolChange
@@ -1518,6 +1475,7 @@ func QueryOnPoolChange(t *Query, limit int, index int, idle int, inUse int) {
 	p.InUse = inUse
 	t.onPoolChange(p)
 }
+
 // Internals: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#internals
 func QueryOnDo(t *Query, c *context.Context, call call) func(attempts int, _ error) {
 	var p QueryDoStartInfo
@@ -1531,6 +1489,7 @@ func QueryOnDo(t *Query, c *context.Context, call call) func(attempts int, _ err
 		res(p)
 	}
 }
+
 // Internals: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#internals
 func QueryOnDoTx(t *Query, c *context.Context, call call) func(attempts int, _ error) {
 	var p QueryDoTxStartInfo
@@ -1544,6 +1503,7 @@ func QueryOnDoTx(t *Query, c *context.Context, call call) func(attempts int, _ e
 		res(p)
 	}
 }
+
 // Internals: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#internals
 func QueryOnExecute(t *Query, c *context.Context, call call, query string) func(error) {
 	var p QueryExecuteStartInfo
@@ -1557,6 +1517,7 @@ func QueryOnExecute(t *Query, c *context.Context, call call, query string) func(
 		res(p)
 	}
 }
+
 // Internals: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#internals
 func QueryOnReadRow(t *Query, c *context.Context, call call, query string) func(error) {
 	var p QueryReadRowStartInfo
@@ -1570,6 +1531,7 @@ func QueryOnReadRow(t *Query, c *context.Context, call call, query string) func(
 		res(p)
 	}
 }
+
 // Internals: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#internals
 func QueryOnReadResultSet(t *Query, c *context.Context, call call, query string) func(error) {
 	var p QueryReadResultSetStartInfo
@@ -1583,6 +1545,7 @@ func QueryOnReadResultSet(t *Query, c *context.Context, call call, query string)
 		res(p)
 	}
 }
+
 // Internals: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#internals
 func QueryOnSessionCreate(t *Query, c *context.Context, call call) func(session querySessionInfo, _ error) {
 	var p QuerySessionCreateStartInfo
@@ -1596,6 +1559,7 @@ func QueryOnSessionCreate(t *Query, c *context.Context, call call) func(session 
 		res(p)
 	}
 }
+
 // Internals: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#internals
 func QueryOnSessionAttach(t *Query, c *context.Context, call call, session querySessionInfo) func(error) {
 	var p QuerySessionAttachStartInfo
@@ -1609,6 +1573,7 @@ func QueryOnSessionAttach(t *Query, c *context.Context, call call, session query
 		res(p)
 	}
 }
+
 // Internals: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#internals
 func QueryOnSessionDelete(t *Query, c *context.Context, call call, session querySessionInfo) func(error) {
 	var p QuerySessionDeleteStartInfo
@@ -1622,6 +1587,7 @@ func QueryOnSessionDelete(t *Query, c *context.Context, call call, session query
 		res(p)
 	}
 }
+
 // Internals: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#internals
 func QueryOnSessionExecute(t *Query, c *context.Context, call call, session querySessionInfo, query string) func(error) {
 	var p QuerySessionExecuteStartInfo
@@ -1636,6 +1602,7 @@ func QueryOnSessionExecute(t *Query, c *context.Context, call call, session quer
 		res(p)
 	}
 }
+
 // Internals: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#internals
 func QueryOnSessionBegin(t *Query, c *context.Context, call call, session querySessionInfo) func(_ error, tx queryTransactionInfo) {
 	var p QuerySessionBeginStartInfo
@@ -1650,6 +1617,7 @@ func QueryOnSessionBegin(t *Query, c *context.Context, call call, session queryS
 		res(p)
 	}
 }
+
 // Internals: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#internals
 func QueryOnTxExecute(t *Query, c *context.Context, call call, session querySessionInfo, tx queryTransactionInfo, query string) func(error) {
 	var p QueryTxExecuteStartInfo
@@ -1665,6 +1633,7 @@ func QueryOnTxExecute(t *Query, c *context.Context, call call, session querySess
 		res(p)
 	}
 }
+
 // Internals: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#internals
 func QueryOnResultNew(t *Query, c *context.Context, call call) func(error) {
 	var p QueryResultNewStartInfo
@@ -1677,19 +1646,7 @@ func QueryOnResultNew(t *Query, c *context.Context, call call) func(error) {
 		res(p)
 	}
 }
-// Internals: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#internals
-func QueryOnResultNextPart(t *Query, c *context.Context, call call) func(stats *Ydb_TableStats.QueryStats, _ error) {
-	var p QueryResultNextPartStartInfo
-	p.Context = c
-	p.Call = call
-	res := t.onResultNextPart(p)
-	return func(stats *Ydb_TableStats.QueryStats, e error) {
-		var p QueryResultNextPartDoneInfo
-		p.Stats = stats
-		p.Error = e
-		res(p)
-	}
-}
+
 // Internals: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#internals
 func QueryOnResultNextResultSet(t *Query, c *context.Context, call call) func(error) {
 	var p QueryResultNextResultSetStartInfo
@@ -1702,6 +1659,7 @@ func QueryOnResultNextResultSet(t *Query, c *context.Context, call call) func(er
 		res(p)
 	}
 }
+
 // Internals: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#internals
 func QueryOnResultClose(t *Query, c *context.Context, call call) func(error) {
 	var p QueryResultCloseStartInfo
@@ -1714,6 +1672,7 @@ func QueryOnResultClose(t *Query, c *context.Context, call call) func(error) {
 		res(p)
 	}
 }
+
 // Internals: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#internals
 func QueryOnResultSetNextRow(t *Query, c *context.Context, call call) func(error) {
 	var p QueryResultSetNextRowStartInfo
@@ -1726,6 +1685,7 @@ func QueryOnResultSetNextRow(t *Query, c *context.Context, call call) func(error
 		res(p)
 	}
 }
+
 // Internals: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#internals
 func QueryOnRowScan(t *Query, c *context.Context, call call) func(error) {
 	var p QueryRowScanStartInfo
@@ -1738,6 +1698,7 @@ func QueryOnRowScan(t *Query, c *context.Context, call call) func(error) {
 		res(p)
 	}
 }
+
 // Internals: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#internals
 func QueryOnRowScanNamed(t *Query, c *context.Context, call call) func(error) {
 	var p QueryRowScanNamedStartInfo
@@ -1750,6 +1711,7 @@ func QueryOnRowScanNamed(t *Query, c *context.Context, call call) func(error) {
 		res(p)
 	}
 }
+
 // Internals: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#internals
 func QueryOnRowScanStruct(t *Query, c *context.Context, call call) func(error) {
 	var p QueryRowScanStructStartInfo